// models/models/HolidaySchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PS Note API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HolidaySchema
 */
export interface HolidaySchema {
    /**
     * 
     * @type {number}
     * @memberof HolidaySchema
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof HolidaySchema
     */
    name: string;
    /**
     * 
     * @type {Date}
     * @memberof HolidaySchema
     */
    date: Date;
}

/**
 * Check if a given object implements the HolidaySchema interface.
 */
export function instanceOfHolidaySchema(value: object): value is HolidaySchema {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('date' in value) || value['date'] === undefined) return false;
    return true;
}

export function HolidaySchemaFromJSON(json: any): HolidaySchema {
    return HolidaySchemaFromJSONTyped(json, false);
}

export function HolidaySchemaFromJSONTyped(json: any, ignoreDiscriminator: boolean): HolidaySchema {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'date': (new Date(json['date'])),
    };
}

export function HolidaySchemaToJSON(value?: HolidaySchema | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'name': value['name'],
        'date': ((value['date']).toISOString().substring(0,10)),
    };
}